YG-182 💡 meta-data에서 일부 내용 상수로 관리하도록 변경
