+++ app/(beforeLogin)/search/page.tsx
@@ -4,7 +4,8 @@ import RealTimeRecommendation from "@/app/_components/userRecommendation/realTim
 import { getPost } from "@/apis/postApi"
 import { Suspense } from "react"
 import { Metadata } from "next"
-import { getSearchMetadata } from "@/constants/metaData"
+import { getSearchMetadata } from "@/utils/metadata.utils"
+
 const SearchClient = dynamic(() => import("./_components/searchClient"), { ssr: false })
 
 // eslint-disable-next-line react-refresh/only-export-components
